+++ sqlglot/expressions.py
@@ -274,12 +274,16 @@ class Expression(metaclass=_Expression):
 
     def set(self, arg_key: str, value: t.Any) -> None:
         """
-        Sets `arg_key` to `value`.
+        Sets arg_key to value.
 
         Args:
-            arg_key (str): name of the expression arg.
+            arg_key: name of the expression arg.
             value: value to set the arg to.
         """
+        if value is None:
+            self.args.pop(arg_key, None)
+            return
+
         self.args[arg_key] = value
         self._set_parent(arg_key, value)
 
@@ -874,11 +878,11 @@ class DerivedTable(Expression):
         return [c.name for c in table_alias.args.get("columns") or []]
 
     @property
-    def selects(self):
+    def selects(self) -> t.List[Expression]:
         return self.this.selects if isinstance(self.this, Subqueryable) else []
 
     @property
-    def named_selects(self):
+    def named_selects(self) -> t.List[str]:
         return [select.output_name for select in self.selects]
 
 
@@ -955,7 +959,7 @@ class Unionable(Expression):
 
 class UDTF(DerivedTable, Unionable):
     @property
-    def selects(self):
+    def selects(self) -> t.List[Expression]:
         alias = self.args.get("alias")
         return alias.columns if alias else []
 
@@ -1572,7 +1576,7 @@ class OnConflict(Expression):
 
 
 class Returning(Expression):
-    arg_types = {"expressions": True}
+    arg_types = {"expressions": True, "into": False}
 
 
 # https://dev.mysql.com/doc/refman/8.0/en/charset-introducer.html
@@ -2190,11 +2194,11 @@ class Subqueryable(Unionable):
         return with_.expressions
 
     @property
-    def selects(self):
+    def selects(self) -> t.List[Expression]:
         raise NotImplementedError("Subqueryable objects must implement `selects`")
 
     @property
-    def named_selects(self):
+    def named_selects(self) -> t.List[str]:
         raise NotImplementedError("Subqueryable objects must implement `named_selects`")
 
     def with_(
@@ -2295,13 +2299,27 @@ class Table(Expression):
         return self.text("catalog")
 
     @property
+    def selects(self) -> t.List[Expression]:
+        return []
+
+    @property
+    def named_selects(self) -> t.List[str]:
+        return []
+
+    @property
     def parts(self) -> t.List[Identifier]:
         """Return the parts of a table in order catalog, db, table."""
-        return [
-            t.cast(Identifier, self.args[part])
-            for part in ("catalog", "db", "this")
-            if self.args.get(part)
-        ]
+        parts: t.List[Identifier] = []
+
+        for arg in ("catalog", "db", "this"):
+            part = self.args.get(arg)
+
+            if isinstance(part, Identifier):
+                parts.append(part)
+            elif isinstance(part, Dot):
+                parts.extend(part.flatten())
+
+        return parts
 
 
 # See the TSQL "Querying data in a system-versioned temporal table" page
@@ -2385,7 +2403,7 @@ class Union(Subqueryable):
         return this
 
     @property
-    def named_selects(self):
+    def named_selects(self) -> t.List[str]:
         return self.this.unnest().named_selects
 
     @property
@@ -2393,7 +2411,7 @@ class Union(Subqueryable):
         return self.this.is_star or self.expression.is_star
 
     @property
-    def selects(self):
+    def selects(self) -> t.List[Expression]:
         return self.this.unnest().selects
 
     @property
@@ -3460,15 +3478,15 @@ class Command(Expression):
 
 
 class Transaction(Expression):
-    arg_types = {"this": False, "modes": False}
+    arg_types = {"this": False, "modes": False, "mark": False}
 
 
 class Commit(Expression):
-    arg_types = {"chain": False}
+    arg_types = {"chain": False, "this": False, "durability": False}
 
 
 class Rollback(Expression):
-    arg_types = {"savepoint": False}
+    arg_types = {"savepoint": False, "this": False}
 
 
 class AlterTable(Expression):
@@ -3509,6 +3527,10 @@ class And(Connector):
 
 
 class Or(Connector):
+    pass
+
+
+class Xor(Connector):
     pass
 
 
@@ -4249,7 +4271,7 @@ class JSONArrayContains(Binary, Predicate, Func):
 
 
 class Least(Func):
-    arg_types = {"expressions": False}
+    arg_types = {"this": True, "expressions": False}
     is_var_len_args = True
 
 
@@ -4342,6 +4364,11 @@ class MD5(Func):
     _sql_names = ["MD5"]
 
 
+# Represents the variant of the MD5 function that returns a binary value
+class MD5Digest(Func):
+    _sql_names = ["MD5_DIGEST"]
+
+
 class Min(AggFunc):
     arg_types = {"this": True, "expressions": False}
     is_var_len_args = True
@@ -4399,6 +4426,7 @@ class RegexpExtract(Func):
         "expression": True,
         "position": False,
         "occurrence": False,
+        "parameters": False,
         "group": False,
     }
 
@@ -5746,7 +5774,9 @@ def table_name(table: Table | str, dialect: DialectType = None) -> str:
         raise ValueError(f"Cannot parse {table}")
 
     return ".".join(
-        part.sql(dialect=dialect) if not SAFE_IDENTIFIER_RE.match(part.name) else part.name
+        part.sql(dialect=dialect, identify=True)
+        if not SAFE_IDENTIFIER_RE.match(part.name)
+        else part.name
         for part in table.parts
     )